fix openvdb.ccp for linux

+++ intern/cycles/render/openvdb.cpp
@@ -558,7 +558,7 @@ void openvdb_load_preprocess(const string& filepath,
 
 	switch(grid_type) {
 		case OPENVDB_GRID_BOOL:
-			return image_load_preprocess<BoolGrid, unsigned long long>(
+			return image_load_preprocess<BoolGrid, unsigned long>(
 			            grid, resolution, min_bound, 1, threshold, use_pad,
 			            sparse_indexes, sparse_size);
 		case OPENVDB_GRID_DOUBLE:
@@ -622,7 +622,7 @@ void openvdb_load_image(const string& filepath,
 	if(make_sparse && use_pad) {
 		switch(grid_type) {
 			case OPENVDB_GRID_BOOL:
-				return image_load_sparse_pad<BoolGrid, unsigned long long>(
+				return image_load_sparse_pad<BoolGrid, unsigned long>(
 				            grid, sparse_indexes, resolution, min_bound, 1,
 				            sparse_size, image);
 			case OPENVDB_GRID_DOUBLE:
@@ -660,7 +660,7 @@ void openvdb_load_image(const string& filepath,
 	else if(make_sparse) {
 		switch(grid_type) {
 			case OPENVDB_GRID_BOOL:
-				return image_load_sparse<BoolGrid, unsigned long long>(
+				return image_load_sparse<BoolGrid, unsigned long>(
 							grid, sparse_indexes, resolution, min_bound, 1, image);
 			case OPENVDB_GRID_DOUBLE:
 				return image_load_sparse<DoubleGrid, double>(
@@ -690,7 +690,7 @@ void openvdb_load_image(const string& filepath,
 	else {
 		switch(grid_type) {
 			case OPENVDB_GRID_BOOL:
-				return image_load_dense<BoolGrid, unsigned long long>(
+				return image_load_dense<BoolGrid, unsigned long>(
 				            grid, resolution, min_bound, 1, image);
 			case OPENVDB_GRID_DOUBLE:
 				return image_load_dense<DoubleGrid, double>(
